requirements.txt
Flask~=3.0.3
requests~=2.32.3
python-dotenv~=1.0.1
WTForms~=3.2.1
redis~=5.2.0
plotly~=5.24.1
dash~=2.18.1
aiogram~=3.13.1
aiohttp~=3.10.10
pydantic~=2.9.2
Hypercorn~=0.17.3